requirements.txt
telethon==1.36.0
aiohttp==3.9.5
python-dotenv==1.0.1
emoji==2.12.1
regex==2024.9.11
discord.py==2.4.0
hyperliquid-python-sdk
